Collect hostboot dump from each SBE
This commit implements the dump create interface to collect the
hostboot dump from each SBE.
A dump entry will be created by requesting to
phosphor-dump-manager, the collected files will be stored
to a temporary path associated with the id.
Scope of this commit is
- Implement xyz::openbmc_project::Dump::Create to accept
the dump parameters such as type of the dump and error log
id.
- Call phosphor dump manager to create an entry
- Collect the dump from each SBE using sbe_dump chipop
- Notify the phosphor dump manager once the dump is
collected.
For Hostboot dump:
- Any error on collecting dump from primary processor
  will fail the dump collection
- Dump collection errors on secondary processors  will
  be ignored.
Tests:
create dump using busctl and make sure the dump is collected
in the path.

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>
Change-Id: Ia65142818ff7723721b78dae19f0d13afc1a33fc

Conflicts:
	dump/dump.mk
	dump/dump_manager.cpp
	dump/dump_manager.hpp