sdevent - cater for releasing linux inode for openpower dumps (#19)
For every dump generated an "anon_inode:[pidfd]" entry is created for the
associated process and the same is not released. When the number of inode
entries reach the maximum count dump request will fail and the dump service
takes a reset with a core dump.

When a dump is requested a child process is spawned to perform dump collection.
Parent process waits on the child process completion using sd-event-add-child
systemd call. sd-event-add-child adds an inode entry which needs to be released
during the callback method called after child process exit using unref but the
same is not done.

Now switching to sdeventplus::source::Child wrapper class for sd-event-add-child
which takes care of releasing the inode fd.

Tested:
1) Succesfully tested creating BMC, hardware, SBE dumps.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I43745a5218b8fe20bcd3ace384fe9d9d4cf197b1